#0028246: A tool is needed to judge the performance of individual machines to figure out what the best performance will be
Added some debugging code. There is now an option of saving the output data in a random folder. We do this because to test the software within muvee, the executable will be left on the server and each pc will call it when it boots up. To prevent each machine from overwriting everyone's results, we enable a flag that creates a folder named after a random number and then stores the results there. I've also disabled the messagebox that used to appear at the end of the test because we will require the server test to run silently.