0013043: muSE-X related changes
Disabled video background if title/credits duration is stretched beyond 5x the original video duration. This would cause the slow-mo factor to be 20% or less. Based on my own test, this is near the threshold where the video is perceived to be jerky. If video background has been disabled, a warning is written to the log.

Fixed a crash issue when video/image does not exist. With this change, the default color background would be used instead. A warning will also be written to the log.