0028912: [ESP] "Trial Version" displayed in the muvee trailer for limit to 60 seconds saving is not localised
Resolved discrepancies in the Spanish locale codes between the app and constructor. The app was using 0x0c0a while the constructor was using 0x040a.

After a quick research on the Internet, I found out that the difference between the locale codes is the sorting order of the Spanish alphabets. The former is generally more preferred. The constructor have been using the latter locale code due to legacy function calls: GetLcidFromRfc1766 and GetRfc1766FromLcid. Since those calls have been deprecated, we can now finally synchronize the Spanish locale code with the app.