0028766: muvee body shows up before title finishes scrolling upwards for Cube Twist style 0028769: Rolling text animation title will overlaps the title video overlay for Uncle Oscar
Fixed an issue which prevented text interval from stopping early. This is required for Uncle Oscar and Techno Buzz title/credits. The problem was traced to an unhandled exception in the last frame of the Text effect.

With that fix in place, we can then reduce the text-display-interval for title and credits when rolling animation is used. This will minimize the overlap between the transition and the rolling text.