0029062: Old text renderer's text shadow is always black color
Reverted changes made in rev 20322 for Journal style only. This prepares the Style module for JPN localization changes for Cogen by removing the code (shadow-color xxx) which is dependent on the latest Constructor module.

I will re-enable the (shadow-color xxx) code after we release the JPN-localized photoElegance stylePack for Cogen.