0029152: [Techno Buzz] Application while previewing muvee on machine that supports new text renderer
Temporary reverted the change introduced in rev 20295. This prepares Adrenaline Rush stylePack to be built using the same set of Constructor and MES modules, which did not include the fix.

This change is for Arcadia_BR2_branch only. The trunk code is not changed. I will reapply this change after we ship Adrenaline Rush stylePack.