#0028313: Video shot using iPhone camera appears in wrong aspect ratio
The decoder correctly rotates iphone protraits video. Some digital cameras however, do not store the video orientation. So, on the long run, muvee reveal will give users the option of rotating the video. I've modified MES to rotate the videos by 90 degrees for now. It's a hack and will support all angles and aspect ratios on the long run. But for now, it's 90 deg at 1.77 AR. To enable it, add C:\\muveeDebug\\RotateVideo.txt