#0029260: Allow videos to be rotated in the media panel (just like we do for pictures) in muvee Reveal
Adding a model view stack in the Cg effect. To cut a long story short, it fixes a video orientation issue in classic vanilla. Cg is the last common point between the two video files that meet during a transition and causes the second video to undergo and extra rotation which is very undesirable.