#0029374: [Cube Twist] Artifacts can be seen in the saved muvee production
When saving muvees on the mac, we were resorting to offscreen rendering with the AGL_OFFSCREEN flag. The problem with this flag is that most hardware renderers do not support it. So the saving was done with software rendering, a very slow process. We never noticed it so far because on the surface, everything seemed to be working fine. However, since muvee reveal express will be shipping with Cube twist and Kinetic styles, some rather severe z-fighting was noticed in the saved production. This happens because the floating point precision of the software renderer is low. Software rendering is, I feel, unacceptable for a software like muvee reveal express.

So I have changed the rendering to the 'hidden window' method that is used in the windows version. Things seem to be ok now. I'll do some more extensive testing now to confirm this.