#0029374: [Cube Twist] Artifacts can be seen in the saved muvee production
Inside ShowTime_RenderEnvBase, we query for the openGL version support. The reason we do this is to set the appropriate texture size based on the openGL support. For machines with an openGL support of < 2.0, we set the texture size to 1k by 1k. If the openGL support is >= 2.0, we set the texture size to 2k by 2k.

To check the openGL version, we need to have a valid openGL context. So far, I had assumed that to be true. But this is not always the case and that caused a crash on Xuiling's machine. This has now been rectified.