0029325: S00567_WinterWonderland Resolved #29534: muvee preview was jerky when aurora sky theme was used
Replaced Aurora Sky intertitle video background with a static image so that the preview is less jerky on low-end machines.

Note: This fixes jerky playback issue for intertitles only. It does not address jerky playback when two or more videos are playing simultaneously. This is a performance issue on low-end systems.