#0029536: Thumbnails in the main UI become grey after saving two muvees
Alright, I found a better way to handle this. I will delete the FBO texture because I do not want any memory leaks from the FBO.

The problem is a conflict in the texture ID. The thumbnails seem to have the texture ID 1. Same for the FBO. When i delete the FBO texture, the thumbnails get deleted as well. So what I can do is generate a dummy texture before generating the actual FBO texture. No conflicts. No memory leaks. Hurray! :)