0030162: [OEM - CoGen] Re-position of captions text for old text renderer
Added TextType to text info hashtable, providing a more elegant way for Kinetic to test for intertitle within the context of muvee-global-effect.

Exposed apply-text-properties in the Caption module.

Deprecate the use of caption text field argument in stylelib2, Captions and TitleAndCredits, when the same content can also be accessed via info.Text.