0030047: [OEM - CoGen] New Animations for Credits
The change in rev 21677 introduced two convenient shorthands: RollingLeftCenterRight and RollingSummaryLeftRight, which are defined in mvtypes.h. But since we are not rebuilding the rest of affected modules dependent on StaticDependencies, I will be reverting the shorthands to the longer form:

TextEffect_RollingLeftCenterRight
=> TextEffect_RollingLeft | TextEffect_Rolling | TextEffect_RollingRight

TextEffect_RollingSummaryLeftRight
=> TextEffect_RollingSummaryLeft | TextEffect_RollingSummaryRight

This change is for BR3 only.