- Updated mvWriter to fix save-ADV-MPEG2 default-baseline.run & save-Burn DVD-baseline.run
Cause of failure: New MC file sink halts when connected with MPEG2 multiplexer, there is no error returned so no way to debug. Tried to tweak the settings but no improvement so temporarily using older version of MC File Sink. This will allow trunk code to work smoothly on MPEG-2 profiles, in the meantime we will  raise this issue to MC for a fix.