Resolved #0035383: Unable to display the string correctly during the title preview
The OpenGL shading language version was lowered from 1.3 to 1.2 on trunk rev 25439, resulting to this machine turning to new text renderer. So we need to revert the version requirement back to 1.3 in order to fix this issue.

Note that this change is currently made for BR5 only. Trunk code remains at version 1.2 for now.