0035299: Cinematic trailers
Fixed resource allocation issue in Trailer04. Textures are now used within the effect times where it is shown. (#0035604)

Resolved an issue with cinematic trailer text not showing up at the last segment if its duration is too long. (#0035523, #0035589)

Worked around an issue in Trailer03 on some machines where it is showing an invalid background texture. This could be an issue with MES. (#0035602)

Bound content enumeration within global effect interval.